rename: mock파일들을 __mock__/responseData 폴더로 이동(Next-Study-habjung#39)
